pool test updated to ts
